Final spagetti update 11/13/14 Rewrite V1
Original Project now can use the Full Conversation Rules. However due to
limitations and design of the orginal project I have started fresh and
captured the same functionality in the importer.py file and
ComplexNPC.py file which now uses JSON and a logical progression of
source management into people into Inform Code. A major refactor of  the
new base is next to abandon dysfunctional aspects and limitations of the
original project